add notifications to cms
