+++ docs/.vitepress/config/data/notifications.json
@@ -1,0 +1,4 @@
+{
+    "banners": [],
+    "notifications": []
+}
