+++ tsconfig.json
@@ -4,7 +4,7 @@
         "docs/.vitepress/**/*.ts"
     ],
     "compilerOptions": {
-        "target": "es2016",
+        "target": "ES2020",
         "module": "ESNext",
         "moduleResolution": "bundler",
         "allowSyntheticDefaultImports": true,
